# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CategoryListResponse": {
        "description": "Every distinct series category.\n\n``count`` is the number of categories, not the number of series; the\nper-category series totals are on the records themselves.",
        "properties": {
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryRecord"
            },
            "title": "Categories",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "count",
          "categories"
        ],
        "title": "CategoryListResponse",
        "type": "object"
      },
      "CategoryRecord": {
        "description": "One series category and how many series carry it.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "series_count": {
            "title": "Series Count",
            "type": "integer"
          }
        },
        "required": [
          "category",
          "series_count"
        ],
        "title": "CategoryRecord",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "The uniform error body for every error this codebase raises (186 D6).\n\nFastAPI's own ``RequestValidationError`` is the one documented exception:\nit keeps its native ``detail`` list, which carries per-field ``loc``/``msg``\nthat a flattened string would lose.",
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EventListResponse": {
        "description": "Events of one series, complete \u2014 never truncated (D4).",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventRecord"
            },
            "title": "Events",
            "type": "array"
          },
          "series_ticker": {
            "title": "Series Ticker",
            "type": "string"
          }
        },
        "required": [
          "series_ticker",
          "count",
          "events"
        ],
        "title": "EventListResponse",
        "type": "object"
      },
      "EventRecord": {
        "description": "A Kalshi event. Field names are the Kalshi/DB names verbatim.",
        "properties": {
          "available_on_brokers": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available On Brokers"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "collateral_return_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collateral Return Type"
          },
          "event_ticker": {
            "title": "Event Ticker",
            "type": "string"
          },
          "first_seen_at": {
            "format": "date-time",
            "title": "First Seen At",
            "type": "string"
          },
          "last_synced_at": {
            "format": "date-time",
            "title": "Last Synced At",
            "type": "string"
          },
          "last_updated_ts": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated Ts"
          },
          "mutually_exclusive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mutually Exclusive"
          },
          "product_metadata": {
            "title": "Product Metadata"
          },
          "series_ticker": {
            "title": "Series Ticker",
            "type": "string"
          },
          "settlement_sources": {
            "title": "Settlement Sources"
          },
          "strike_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strike Date"
          },
          "strike_period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strike Period"
          },
          "sub_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Title"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "event_ticker",
          "series_ticker",
          "title",
          "sub_title",
          "category",
          "mutually_exclusive",
          "strike_date",
          "strike_period",
          "collateral_return_type",
          "available_on_brokers",
          "settlement_sources",
          "product_metadata",
          "last_updated_ts",
          "first_seen_at",
          "last_synced_at"
        ],
        "title": "EventRecord",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MarketListResponse": {
        "description": "Markets of one event, complete \u2014 never truncated (D4).",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "event_ticker": {
            "title": "Event Ticker",
            "type": "string"
          },
          "markets": {
            "items": {
              "$ref": "#/components/schemas/MarketRecord"
            },
            "title": "Markets",
            "type": "array"
          }
        },
        "required": [
          "event_ticker",
          "count",
          "markets"
        ],
        "title": "MarketListResponse",
        "type": "object"
      },
      "MarketRecord": {
        "description": "A Kalshi market: every typed column except ``raw`` (D2).\n\nThe lifecycle / settlement / economics grouping is for readability; the\nfield names inside each group are the DB/Kalshi names verbatim, so\nKalshi's own field documentation applies to them unchanged.",
        "properties": {
          "economics": {
            "$ref": "#/components/schemas/MarketEconomics"
          },
          "event_ticker": {
            "title": "Event Ticker",
            "type": "string"
          },
          "first_seen_at": {
            "format": "date-time",
            "title": "First Seen At",
            "type": "string"
          },
          "is_provisional": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Provisional"
          },
          "last_synced_at": {
            "format": "date-time",
            "title": "Last Synced At",
            "type": "string"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/MarketLifecycle"
          },
          "market_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Type"
          },
          "mve_collection_ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mve Collection Ticker"
          },
          "no_sub_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "No Sub Title"
          },
          "price_level_structure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Level Structure"
          },
          "rules_primary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rules Primary"
          },
          "rules_secondary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rules Secondary"
          },
          "settlement": {
            "$ref": "#/components/schemas/MarketSettlement"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "strike_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strike Type"
          },
          "subtitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "yes_sub_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yes Sub Title"
          }
        },
        "required": [
          "ticker",
          "event_ticker",
          "market_type",
          "status",
          "title",
          "subtitle",
          "yes_sub_title",
          "no_sub_title",
          "rules_primary",
          "rules_secondary",
          "strike_type",
          "price_level_structure",
          "is_provisional",
          "mve_collection_ticker",
          "first_seen_at",
          "last_synced_at",
          "lifecycle",
          "settlement",
          "economics"
        ],
        "title": "MarketRecord",
        "type": "object"
      },
...
      "SeriesListResponse": {
        "description": "Series matching the requested filter, complete \u2014 never truncated (D4).",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/SeriesRecord"
            },
            "title": "Series",
            "type": "array"
          }
        },
        "required": [
          "count",
          "series"
        ],
        "title": "SeriesListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Data serving API for OHLCV bars, symbol metadata, gap status, the Kalshi prediction-market catalog and time series, and operations coverage: pass state, source freshness and provider credits.",
    "title": "Manta Trading API",
    "version": "0.16.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/kalshi/categories": {
      "get": {
        "description": "Return every series category with the number of series carrying it.\n\nCategories are free text Kalshi assigns, so this is how a client learns\nwhich values the series list's ``category`` filter accepts. The series\ncounts double as a size map for planning those calls.",
        "operationId": "list_categories_api_v1_kalshi_categories_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The database cancelled the query at the server's statement timeout. Narrow the requested range or use a coarser granularity."
          }
        },
        "summary": "List Categories"
      }
    },
    "/api/v1/kalshi/events/{event_ticker}": {
      "get": {
        "description": "Return one event by its event ticker.",
        "operationId": "get_event_api_v1_kalshi_events__event_ticker__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_ticker",
            "required": true,
            "schema": {
              "title": "Event Ticker",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The database cancelled the query at the server's statement timeout. Narrow the requested range or use a coarser granularity."
          }
        },
        "summary": "Get Event"
      }
    },
    "/api/v1/kalshi/events/{event_ticker}/markets": {
      "get": {
        "description": "Return the markets of one event, optionally filtered by status.\n\n``status`` accepts a comma-separated list of market statuses; omit it for\nevery status. Naming a value the column cannot hold is an error rather\nthan an empty result.",
        "operationId": "list_markets_api_v1_kalshi_events__event_ticker__markets_get",
        "parameters": [
          {
            "in": "path",
            "name": "event_ticker",
            "required": true,
            "schema": {
              "title": "Event Ticker",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The database cancelled the query at the server's statement timeout. Narrow the requested range or use a coarser granularity."
          }
        },
        "summary": "List Markets"
      }
    },
    "/api/v1/kalshi/markets/{ticker}": {
      "get": {
        "description": "Return one market by its ticker.",
        "operationId": "get_market_api_v1_kalshi_markets__ticker__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "title": "Ticker",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The database cancelled the query at the server's statement timeout. Narrow the requested range or use a coarser granularity."
          }
        },
        "summary": "Get Market"
      }
    },
...
    "/api/v1/kalshi/series": {
      "get": {
        "description": "Return series matching an optional category and ticker-prefix search.\n\n``category`` is an exact match; ``search`` matches the start of a ticker.\nThe response is complete or refused \u2014 there is no pagination and no\ntruncation, so a 200 always carries every matching row.",
        "operationId": "list_series_api_v1_kalshi_series_get",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The database cancelled the query at the server's statement timeout. Narrow the requested range or use a coarser granularity."
          }
        },
        "summary": "List Series"
      }
    },
    "/api/v1/kalshi/series/{ticker}": {
      "get": {
        "description": "Return one series by its ticker.",
        "operationId": "get_series_api_v1_kalshi_series__ticker__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "title": "Ticker",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The database cancelled the query at the server's statement timeout. Narrow the requested range or use a coarser granularity."
          }
        },
        "summary": "Get Series"
      }
    },
    "/api/v1/kalshi/series/{ticker}/events": {
      "get": {
        "description": "Return the events of one series, optionally bounded by strike date.\n\n``strike_from`` and ``strike_to`` are inclusive. An event with no strike\ndate is excluded whenever either bound is given, since a null cannot\nsatisfy a comparison.",
        "operationId": "list_events_api_v1_kalshi_series__ticker__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "title": "Ticker",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "strike_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Strike From"
            }
          },
          {
            "in": "query",
            "name": "strike_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Strike To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The database cancelled the query at the server's statement timeout. Narrow the requested range or use a coarser granularity."
          }
        },
        "summary": "List Events"
      }
    },
...
  }
}